services/ai-agent/requirements.txt
minio==7.2.0
pypdf==3.17.4
google-generativeai
confluent-kafka
protobuf
qdrant-client
